fix(modal): refactor
