+++ packages/react/src/components/modal/dialog/modal-dialog.test.tsx
@@ -1,10 +1,10 @@
 import { fireEvent, RenderResult } from '@testing-library/react';
-import { doNothing } from '../../test-utils/callbacks';
-import { getByTestId as enzymeGetByTestId } from '../../test-utils/enzyme-selectors';
-import { mountWithProviders, renderPortalWithProviders } from '../../test-utils/renderer';
-import { DeviceType } from '../device-context-provider/device-context-provider';
+import { doNothing } from '../../../test-utils/callbacks';
+import { getByTestId as enzymeGetByTestId } from '../../../test-utils/enzyme-selectors';
+import { mountWithProviders, renderPortalWithProviders } from '../../../test-utils/renderer';
+import { DeviceType } from '../../device-context-provider/device-context-provider';
 import { ModalDialog } from './modal-dialog';
-import { IconName } from '../icon/icon';
+import { IconName } from '../../icon/icon';
 import { DialogType, ModalDialogProps } from './types';
 
 type ModalDialogPropsLite = Omit<ModalDialogProps, 'ariaDescribedby' | 'ariaHideApp' | 'onRequestClose' | 'title'>;
