+++ packages/react/src/components/modal/index.ts
@@ -1,2 +1,2 @@
 export { Modal } from './modal';
-export { ModalDialog } from './modal-dialog';
+export { ModalDialog } from './dialog/modal-dialog';
